requirements.txt
Flask
numpy
pandas
Pillow
pickle5
scikit-learn